# docs/integrations/api/openapi.yaml
openapi: 3.0.3
info:
  title: Landslide Monitoring API (v2)
  version: 0.1.0
  description: |
    本文件是 API 契约的机器可读入口（契约优先），与 `docs/integrations/api/*.md` 保持一致。
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /gps/baselines:
    get:
      tags: [gps]
      summary: List GPS baselines
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [list, pagination]
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/GpsBaseline"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /gps/baselines/{deviceId}:
    get:
      tags: [gps]
      summary: Get GPS baseline by deviceId
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GpsBaseline"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

    put:
      tags: [gps]
      summary: Upsert GPS baseline (manual)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertGpsBaselineRequest"
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

    delete:
      tags: [gps]
      summary: Delete GPS baseline
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  responses:
    SuccessResponse:
      description: ok
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    ErrorResponse:
      description: error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedResponse:
      description: unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenResponse:
      description: forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundResponse:
      description: not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerErrorResponse:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessResponse:
      type: object
      required: [success, code, message, timestamp, traceId]
      properties:
        success: { type: boolean, enum: [true] }
        code: { type: integer, example: 200 }
        message: { type: string, example: ok }
        data: {}
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00Z"
        traceId:
          type: string
          example: "req_01J..."
    ErrorResponse:
      type: object
      required: [success, code, message, timestamp, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code: { type: integer, example: 400 }
        message: { type: string, example: 参数错误 }
        error:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00Z"
        traceId:
          type: string
          example: "req_01J..."
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    GpsBaselinePoint:
      type: object
      required: [latitude, longitude]
      additionalProperties: false
      properties:
        latitude: { type: number }
        longitude: { type: number }
        altitude: { type: number }
        positionAccuracyMeters: { type: number }
        satelliteCount: { type: integer, minimum: 1 }
        notes: { type: string }
    GpsBaseline:
      type: object
      required: [deviceId, deviceName, method, baseline, computedAt, updatedAt]
      properties:
        deviceId: { type: string, format: uuid }
        deviceName: { type: string }
        stationId: { type: string, format: uuid, nullable: true }
        method: { type: string, enum: [auto, manual] }
        pointsCount: { type: integer, nullable: true }
        baseline:
          $ref: "#/components/schemas/GpsBaselinePoint"
        computedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpsertGpsBaselineRequest:
      type: object
      required: [baseline]
      additionalProperties: false
      properties:
        method: { type: string, enum: [auto, manual] }
        pointsCount: { type: integer, minimum: 1 }
        baseline:
          $ref: "#/components/schemas/GpsBaselinePoint"
...